Move channel classes into sub-package
