+++ src/main/java/de/jcm/discordgamesdk/Core.java
@@ -3,6 +3,9 @@ package de.jcm.discordgamesdk;
 import com.google.gson.Gson;
 import de.jcm.discordgamesdk.impl.Error;
 import de.jcm.discordgamesdk.impl.*;
+import de.jcm.discordgamesdk.impl.channel.DiscordChannel;
+import de.jcm.discordgamesdk.impl.channel.UnixDiscordChannel;
+import de.jcm.discordgamesdk.impl.channel.WindowsDiscordChannel;
 import de.jcm.discordgamesdk.impl.commands.Subscribe;
 import de.jcm.discordgamesdk.impl.events.OverlayUpdateEvent;
 import de.jcm.discordgamesdk.impl.events.VoiceSettingsUpdate2Event;
