+++ src/main/java/de/jcm/discordgamesdk/impl/channel/UnixDiscordChannel.java
@@ -1,4 +1,4 @@
-package de.jcm.discordgamesdk;
+package de.jcm.discordgamesdk.impl.channel;
 
 import java.io.File;
 import java.io.IOException;
